external deanotation

+++ inspections/impl/com/intellij/codeInspection/RemoveAnnotationQuickFix.java
@@ -2,9 +2,11 @@ package com.intellij.codeInspection;
 
 import com.intellij.codeInsight.CodeInsightBundle;
 import com.intellij.codeInsight.CodeInsightUtil;
+import com.intellij.codeInsight.ExternalAnnotationsManager;
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiAnnotation;
+import com.intellij.psi.PsiModifierListOwner;
 import com.intellij.util.IncorrectOperationException;
 import org.jetbrains.annotations.NotNull;
 
@@ -14,9 +16,11 @@ import org.jetbrains.annotations.NotNull;
 public class RemoveAnnotationQuickFix implements LocalQuickFix {
   private static final Logger LOG = Logger.getInstance("com.intellij.codeInsight.i18n.AnnotateNonNlsQuickfix");
   private final PsiAnnotation myAnnotation;
+  private final PsiModifierListOwner myListOwner;
 
-  public RemoveAnnotationQuickFix(PsiAnnotation annotation) {
+  public RemoveAnnotationQuickFix(PsiAnnotation annotation, final PsiModifierListOwner listOwner) {
     myAnnotation = annotation;
+    myListOwner = listOwner;
   }
 
   @NotNull
@@ -25,13 +29,16 @@ public class RemoveAnnotationQuickFix implements LocalQuickFix {
   }
 
   public void applyFix(@NotNull Project project, @NotNull ProblemDescriptor descriptor) {
-    if (!CodeInsightUtil.preparePsiElementForWrite(myAnnotation)) return;
-
-    try {
-      myAnnotation.delete();
-    }
-    catch (IncorrectOperationException e) {
-      LOG.error(e);
+    if (myAnnotation.isPhysical()) {
+      try {
+        if (!CodeInsightUtil.preparePsiElementForWrite(myAnnotation)) return;
+        myAnnotation.delete();
+      }
+      catch (IncorrectOperationException e) {
+        LOG.error(e);
+      }
+    } else {
+      ExternalAnnotationsManager.getInstance().deannotate(myListOwner, myAnnotation.getQualifiedName());
     }
   }
